✨(frontend) extract useRoomSetup facade from Conference
Extract room options, connection warm-up, and server URL logic
into a useRoomSetup hook. Reduces Conference.tsx from 7+ concerns
to 4, applying Facade pattern.

Closes #52
