+++ CHANGELOG.md
@@ -7,9 +7,10 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- ✨(frontend) extract PanelToggleButton to eliminate duplicated
-  panel toggle structure (issue #54)
-- ✨(frontend) replace custom debounce with useDebounceValue from usehooks-ts
+- ✨(rooms) introduce facade hooks for useRoomContext (issue #55)
+- ✨(frontend) extract PanelToggleButton (issue #54)
+- ✨(frontend) replace custom debounce with usehooks-ts
+- ✨(frontend) extract useRoomSetup facade from Conference
 - Deduplicated posthog ingress version and backend logic via
   named helpers in `_helpers.tpl` (issue #24)
   - Deduplicated ingress admin version and backend logic via
